+++ src/unity-examples/item/Item_AppleSO.cs
@@ -9,7 +9,7 @@ namespace WWWisky.inventory.unity.examples
     [CreateAssetMenu]
     public class Item_AppleSO : ItemSO
     {
-        [SerializeField] private float Cost;
+        [SerializeField] private int Cost;
 
         /// <summary>
         /// 
@@ -29,7 +29,7 @@ namespace WWWisky.inventory.unity.examples
     {
         public int StackSize { get; } = 5;
         public bool IsExpendedOnUse { get; } = true;
-        public float Cost { get; }
+        public int Cost { get; }
 
 
         /// <summary>
@@ -38,7 +38,7 @@ namespace WWWisky.inventory.unity.examples
         /// <param name="id"></param>
         /// <param name="name"></param>
         /// <param name="cost"></param>
-        public Item_Apple(string id, string name, float cost) : base(id, name, ItemType.Consumable)
+        public Item_Apple(string id, string name, int cost) : base(id, name, ItemType.Consumable)
         {
             Cost = cost;
         }
